requirements.txt
# ============================================
# Multi-Agent E-Commerce System Requirements
# ============================================
# Python 3.8+ required

# ============================================
# Core Dependencies (Required)
# ============================================

google-generativeai==0.3.1
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0

# ============================================
# Optional Dependencies
# ============================================

# For advanced similarity search (if scaling)
# scikit-learn>=1.3.0

# For vector databases (if scaling to large datasets)
# faiss-cpu>=1.7.4

# ============================================
# Development Tools (Optional)
# ============================================

# Jupyter notebook support
# jupyter>=1.0.0
# ipython>=8.0.0
# notebook>=7.0.0

# Code formatting
# black>=23.0.0
# autopep8>=2.0.0

# Linting
# pylint>=2.17.0
# flake8>=6.0.0

# Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0

# ============================================
# Installation Instructions
# ============================================

# Install only required dependencies:
# pip install -r requirements.txt

# Install with development tools:
# pip install -r requirements.txt jupyter black pytest

# ============================================
# Version Notes
# ============================================

# google-generativeai==0.3.1
#   - Stable version with free tier support
#   - Includes all Gemini models (Flash, Pro)
#   - Compatible with rate limiting

# numpy>=1.24.0
#   - Used for similarity calculations
#   - Vector operations in Agent 1

# pandas>=2.0.0
#   - CSV file loading
#   - Data validation
#   - Data merging (products + inventory)

# python-dotenv>=1.0.0
#   - Loads .env file
#   - Manages API keys securely
#   - Environment variable handling